+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Development
 
+### Added
+
+* Integration test harness to start DLM services (via Docker Compose).
+* Start of integration tests for directory-watcher and kafka-watcher.
+
 ### Changed
 
 * Kafka Docker image switched from bitnami/kafka to apache/kafka